Outsourced robot-related logic into additional robot controller
